requirements.txt
beautifulsoup4
numpy
pandas
plotly-express
pytrends
requests
tqdm
